requirements.txt
absl-py==2.2.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
faster-coco-eval==1.6.5
filelock==3.18.0
fire==0.7.0
fonttools==4.57.0
fsspec==2025.3.2
grpcio==1.71.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
narwhals==1.37.1
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.10.0.84
packaging==25.0
pandas==2.2.3
pillow==11.2.1
plotly==6.0.1
protobuf==6.30.2
psutil==7.0.0
py-cpuinfo==9.0.0
pybboxes==0.1.6
pycocotools==2.0.8
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sahi==0.11.22
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.0
shapely==2.1.0
six==1.17.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminaltables==3.1.10
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
typing_extensions==4.13.2
tzdata==2025.2
-e git+https://github.com/ultralytics/ultralytics.git@e1d742500fbfc24a08c4eb664d0111eaffc8827a#egg=ultralytics
ultralytics-thop==2.0.14
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
